+++ core/api/typescript/src/lib/services/container_config.ts
@@ -12,6 +12,8 @@ export class ContainerConfig {
     constructor(
         public readonly image: string,
         public readonly usedPorts: Map<string, PortSpec>,
+        // TODO REMOVE
+        public readonly oldFilesArtifactMountpoints: Map<FilesArtifactID, string>,
         public readonly filesArtifactMountpoints: Map<FilesArtifactID, string>,
         public readonly entrypointOverrideArgs: string[],
         public readonly cmdOverrideArgs: string[],
@@ -29,6 +31,7 @@ export class ContainerConfig {
 export class ContainerConfigBuilder {
     private readonly image: string;
     private usedPorts: Map<string, PortSpec>;
+    private oldFilesArtifactMountpoints: Map<FilesArtifactID, string>;
     private filesArtifactMountpoints: Map<FilesArtifactID, string>;
     private entrypointOverrideArgs: string[];
 	private cmdOverrideArgs: string[];
@@ -37,6 +40,7 @@ export class ContainerConfigBuilder {
     constructor (image: string) {
         this.image = image;
         this.usedPorts = new Map();
+        this.oldFilesArtifactMountpoints = new Map();
         this.filesArtifactMountpoints = new Map();
         this.entrypointOverrideArgs = [];
         this.cmdOverrideArgs = [];
@@ -49,6 +53,11 @@ export class ContainerConfigBuilder {
     }
 
     public withFilesArtifacts(filesArtifactMountpoints: Map<FilesArtifactID, string>): ContainerConfigBuilder {
+        this.oldFilesArtifactMountpoints = filesArtifactMountpoints;
+        return this;
+    }
+
+    public withFiles(filesArtifactMountpoints: Map<FilesArtifactID, string>): ContainerConfigBuilder {
         this.filesArtifactMountpoints = filesArtifactMountpoints;
         return this;
     }
@@ -72,6 +81,7 @@ export class ContainerConfigBuilder {
         return new ContainerConfig(
             this.image,
             this.usedPorts,
+            this.oldFilesArtifactMountpoints,
             this.filesArtifactMountpoints,
             this.entrypointOverrideArgs,
             this.cmdOverrideArgs,